+++ .gitignore
@@ -1,1 +1,2 @@
+/build/
 /vendor/
